Group abstract classes into a separate directory.
