+++ CarRental.PresentationLayer/Models/Car/Abstractions/AbstractWheels.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Drawing;
 using CarRental.Data.Enums;
 using CarRental.Data.Interfaces;
 
@@ -12,9 +11,9 @@ using CarRental.Data.Interfaces;
 // STUDENT: PARKHOMENKO YAROSLAV
 // DATE: 03-JUN-2024
 
-namespace CarRental.Data.Models.Car;
+namespace CarRental.Data.Models.Car.Abstractions;
 
-public abstract class AbstractLights : IComponent
+public abstract class AbstractWheels : IComponent
 {
     // THE PURPOSE OF THE CLASS:
     // // A CAR'S COMPONENT ABSTRACTION.
@@ -22,7 +21,8 @@ public abstract class AbstractLights : IComponent
     // PROPERTIES
 
     public abstract required string SerialNumber { get; init; }
-    public abstract PowerComponent Power { get; init; }
-    public abstract KnownColor Color { get; init; }
-    public abstract required ComponentStatus Status { get; set; }
+    public abstract MaterialWheel Material { get; init; }
+    public abstract int Size { get; init; }
+    public abstract required TypeTire Tire { get; set; }
+    public abstract ComponentStatus Status { get; set; }
 }
