+++ CarRental.PresentationLayer/Models/Car/Engine.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Diagnostics.CodeAnalysis;
 using CarRental.Data.Enums;
+using CarRental.Data.Models.Car.Abstractions;
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
 // HOMEWORK: "ДЗ 4. Extension methods & Record type."
